requirements.txt
numpy
tensorflow
keras
tqdm
streamlit
scikit-learn
opencv-python
pillow
gunicorn
face_recognition
